@@ -1,4 +1,3 @@
-import 'package:flare_dart/math/mat2d.dart';
 import 'package:flare_flutter/flare.dart';
 import 'package:flare_flutter/flare_actor.dart';
 import 'package:flare_flutter/flare_controller.dart';
@@ -8,18 +7,18 @@ import 'package:flutter_ecommerce_app/src/utils/UBScaffold/page_state.dart';
 import 'buttons_button_sheet.dart';
 
 class ConfirmationBottomSheet extends StatefulWidget {
-  final String title;
-  final String message;
-  final String confirmMessage;
+  final String? title;
+  final String? message;
+  final String? confirmMessage;
   final bool confirmIsLoading;
-  final String cancelMessage;
-  final Function confirmAction;
-  final Function cancelAction;
-  final String icon;
-  final String flare;
+  final String? cancelMessage;
+  final Function? confirmAction;
+  final Function? cancelAction;
+  final String? icon;
+  final String? flare;
 
   const ConfirmationBottomSheet({
-    Key key,
+    Key? key,
     this.title,
     this.message,
     this.confirmMessage,
@@ -38,8 +37,8 @@ class ConfirmationBottomSheet extends StatefulWidget {
 
 class GRConfirmationBottomSheetState extends State<ConfirmationBottomSheet>
     with FlareController {
-  PageState actionState;
-  FlutterColorFill _fill;
+  PageState? actionState;
+  FlutterColorFill? _fill;
 
   @override
   Widget build(BuildContext context) {
@@ -52,7 +51,7 @@ class GRConfirmationBottomSheetState extends State<ConfirmationBottomSheet>
               crossAxisAlignment: CrossAxisAlignment.center,
               children: [
                 SizedBox(height: 28),
-                if (widget.icon != null) Image.asset(widget.icon),
+                if (widget.icon != null) Image.asset(widget.icon ?? ""),
                 if (widget.flare != null)
                   Container(
                     height: 100,
@@ -92,13 +91,19 @@ class GRConfirmationBottomSheetState extends State<ConfirmationBottomSheet>
                   ),
                 SizedBox(height: 24),
                 ButtonsButtonSheet(
-                  cancelMessage: widget.cancelMessage,
-                  confirmMessage: widget.confirmMessage,
+                  cancelMessage: widget.cancelMessage ?? "",
+                  confirmMessage: widget.confirmMessage ?? "",
                   confirmIsLoading: actionState != null
                       ? actionState == PageState.loading
                       : widget.confirmIsLoading,
-                  confirmAction: widget.confirmAction,
-                  cancelAction: widget.cancelAction,
+                  confirmAction: () {
+                    Navigator.of(context).pop();
+                    if (widget.confirmAction != null) widget.confirmAction!();
+                  },
+                  cancelAction: () {
+                    Navigator.of(context).pop();
+                    if (widget.cancelAction != null) widget.cancelAction!();
+                  },
                 ),
                 SizedBox(height: 22),
               ],
@@ -111,15 +116,15 @@ class GRConfirmationBottomSheetState extends State<ConfirmationBottomSheet>
 
   @override
   void initialize(FlutterActorArtboard artboard) {
-    FlutterActorShape shape = artboard.getNode("circle_bg");
-    _fill = shape?.fill as FlutterColorFill;
+    var shape = artboard.getNode("circle_bg");
+    // _fill = shape?.fill as FlutterColorFill;
   }
 
   @override
   bool advance(FlutterActorArtboard artboard, double elapsed) {
     Color nextColor = Colors.white;
     if (_fill != null) {
-      _fill.uiColor = nextColor;
+      // _fill.uiColor = nextColor;
     }
     return false;
   }
