@@ -1,23 +1,23 @@
 import 'package:flutter/material.dart';
 
 class UBAppBar extends StatefulWidget implements PreferredSizeWidget {
-  final Key key;
-  final Widget leading;
-  final bool automaticallyImplyLeading;
-  final Widget title;
-  final List<Widget> actions;
-  final Widget flexibleSpace;
-  final PreferredSizeWidget bottom;
-  final double elevation;
-  final Color backgroundColor;
-  final Brightness brightness;
-  final IconThemeData iconTheme;
-  final TextTheme textTheme;
-  final bool primary;
-  final bool centerTitle;
-  final double titleSpacing;
-  final double toolbarOpacity;
-  final double bottomOpacity;
+  final Key? key;
+  final Widget? leading;
+  final bool? automaticallyImplyLeading;
+  final Widget? title;
+  final List<Widget>? actions;
+  final Widget? flexibleSpace;
+  final PreferredSizeWidget? bottom;
+  final double? elevation;
+  final Color? backgroundColor;
+  final Brightness? brightness;
+  final IconThemeData? iconTheme;
+  final TextTheme? textTheme;
+  final bool? primary;
+  final bool? centerTitle;
+  final double? titleSpacing;
+  final double? toolbarOpacity;
+  final double? bottomOpacity;
 
   UBAppBar({
     this.key,
@@ -45,7 +45,7 @@ class UBAppBar extends StatefulWidget implements PreferredSizeWidget {
 
   @override
   Size get preferredSize =>
-      Size.fromHeight(kToolbarHeight + (bottom?.preferredSize?.height ?? 0.0));
+      Size.fromHeight(kToolbarHeight + (bottom?.preferredSize.height ?? 0.0));
 }
 
 class _UBAppBarState extends State<UBAppBar> {
@@ -53,11 +53,11 @@ class _UBAppBarState extends State<UBAppBar> {
   Widget build(BuildContext context) {
     widget.createElement().state.setState(() {});
 
-    Color color;
+    Color? color;
     return AppBar(
       key: widget.key,
       leading: widget.leading,
-      automaticallyImplyLeading: widget.automaticallyImplyLeading,
+      automaticallyImplyLeading: widget.automaticallyImplyLeading!,
       title: widget.title,
       actions: widget.actions,
       flexibleSpace: widget.flexibleSpace,
@@ -67,11 +67,11 @@ class _UBAppBarState extends State<UBAppBar> {
       brightness: widget.brightness,
       iconTheme: widget.iconTheme,
       textTheme: widget.textTheme,
-      primary: widget.primary,
+      primary: widget.primary ?? false,
       centerTitle: widget.centerTitle,
       titleSpacing: widget.titleSpacing,
-      toolbarOpacity: widget.toolbarOpacity,
-      bottomOpacity: widget.bottomOpacity,
+      toolbarOpacity: widget.toolbarOpacity ?? 1.0,
+      bottomOpacity: widget.bottomOpacity ?? 1.0,
     );
   }
 }
