@@ -6,14 +6,14 @@ import 'error_switcher.dart';
 import 'loader.dart';
 
 class UBPageStateWidget extends StatelessWidget {
-  final PageState pageState;
-  final Widget loadingWidget;
-  final WidgetBuilder builder;
-  final WidgetBuilder noDataBuilder;
-  final String textUnderLoader;
-  final Function onRetry;
+  final PageState? pageState;
+  final Widget? loadingWidget;
+  final WidgetBuilder? builder;
+  final WidgetBuilder? noDataBuilder;
+  final String? textUnderLoader;
+  final Function? onRetry;
   final dynamic error;
-  final String noDataMessage;
+  final String? noDataMessage;
 
   UBPageStateWidget({
     this.pageState,
@@ -34,19 +34,19 @@ class UBPageStateWidget extends StatelessWidget {
         pageBody = loadingWidget ?? Loader();
         break;
       case PageState.loaded:
-        if (builder != null) pageBody = Builder(builder: builder);
+        if (builder != null) pageBody = Builder(builder: builder!);
         break;
       case PageState.error:
         pageBody = ErrorSwitcher(
           message: Localization.of(context, 'network_error'),
-          onRetry: onRetry,
+          onRetry: () => onRetry,
           error: error,
         );
         break;
       case PageState.noData:
         if (noDataBuilder != null)
-          pageBody = Builder(builder: noDataBuilder);
-        else if (noDataMessage != null) pageBody = NoData(noDataMessage);
+          pageBody = Builder(builder: noDataBuilder!);
+        else if (noDataMessage != null) pageBody = NoData(noDataMessage!);
         break;
     }
 
@@ -58,11 +58,11 @@ class NoData extends StatelessWidget {
   NoData(
     this.message, {
     this.child,
-    Key key,
+    Key? key,
     this.size = 18,
   });
 
-  final Widget child;
+  final Widget? child;
   final String message;
   final double size;
 
